Fix the problem with Refresh
