+++ src/index.html
@@ -15,11 +15,13 @@
             <div id="score-container">
                 <div id="scoretitle-container">
                     <h2 class="score-title">Recent scores</h2>
-                    <button class="refresh-button">Refresh</button>
+                    <button class="refresh-button" id="refresh-button">Refresh</button>
                 </div>
-                <ul id="score-list" class="score-list">
+                <div id="">
+                    <ul id="score-list" class="score-list">
                     
-                </ul>
+                    </ul>
+                </div>
             </div>
             <div id="form">
                 <h2 id="form-title">Add your score</h2>
