+++ src/modules/output.js
@@ -1,19 +1,24 @@
-import {aplication} from './variables.js';
+import { aplication } from "./variables";
 
-fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/777/scores/')
-    .then((response) => response.json())
-    .then((json) => showRefresh(json))
-    .catch((err) => console.log(err));
-  
-  document.querySelector('.refresh-button').addEventListener('click', () => {
-    location.reload();
-  });
+const showRefresh = (e) => {
+  for (let i = 0; i < e.result.length; i++) {
+    const li = document.createElement('li');
+    li.innerHTML = `${e.result[i].user} ${e.result[i].score}`;
+    aplication.appendChild(li);
+  }
+};
 
-const showRefresh = (val) => {
-    for (let i = 0; i < val.result.length; i++) {
-      const li = document.createElement('li');
-      li.innerHTML = val.result[i].user + val.result[i].score;
-      aplication.appendChild(li);
-    }
-  };
-  
+const remov = async () => {
+  await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/2225/scores/')
+  .then((response) => response.json())
+  .then((json) => showRefresh(json))
+  .catch((err) => console.log(err));
+}
+remov();
+
+document.getElementById('refresh-button').addEventListener('click', () => {
+  while(aplication.hasChildNodes()) {
+    aplication.removeChild(aplication.firstChild)
+  }
+  remov();
+});
